Refactor plugin functionality
